Add a set selection interface based on the periodic table (#1237)
* Add two queries related to periodic elements

`getActiveElements` retrieves all elements that occur in a published set.
`getSetHeaderElements` retrieves set headers for sets that contain an
element that occurs in the given selection.

* Initial implementation of the `PeriodicTable` component

Note that the current implementation is broken due to faulty
usage of a server action.

* Add `organization` key to `PeriodicSearchResult`

* Add `useEffect` hook to call server function

Add `onChange` callback to communicate search results.

* Render search results in a table

Each row is clickable to go to the inspection page of the
corresponding dataset.
Most functionality was moved to a `PageClient` client
component, such that it can keep state.

* Add `Fieldset` for the search results

* Tweak layout

* Point navbar to new periodic table page

* Cleanup database package

* Remove `scat-css` page

* Add function to list public sets

* Use `listSets` function to fix test in `delete-set`

* Fix `scat-css` endpoint by using `listSets` function

* Fix REUSE compliance

* Add missing dependency to `useEffect` hook

* Opt-out of SSG for `cs-set` page

* Remove unused dependencies

* Remove unused modules

* Test the periodic select component in end to end tests

* Do not truncate the `Element` collection

* Return the set name in `listSets` instead of the org name

* Add unit tests for element queries

+++ packages/database/src/lxcat-database.ts
@@ -88,13 +88,13 @@ import {
   activeSetOfArchivedSet,
   byId as setById,
   getItemIdsInSet,
-  search as searchSet,
-  searchContributors,
-  searchFacets,
+  listSets,
   setHistory,
-  stateChoices,
-  tagChoices,
 } from "./css/queries/public.js";
+import {
+  getActiveElements,
+  getSetHeaderByElements,
+} from "./elements/queries.js";
 import { setupCompositionCollections } from "./setup/composition.js";
 import { setupCrossSectionCollections } from "./setup/cs.js";
 import { setupDatabase } from "./setup/db.js";
@@ -233,8 +233,8 @@ export class LXCatDatabase {
   public async truncateNonUserCollections() {
     const collections = await this.db.collections(true);
     for (const c of collections) {
-      console.log(`Truncating ${c.name}`);
-      if (c.name !== "users") {
+      if (c.name !== "users" && c.name !== "Element") {
+        console.log(`Truncating ${c.name}`);
         await c.truncate();
       }
     }
@@ -309,6 +309,10 @@ export class LXCatDatabase {
   public getTopLevelSpecies = getTopLevelSpecies;
   public getSpeciesChildren = getSpeciesChildren;
 
+  // elements
+  public getActiveElements = getActiveElements;
+  public getSetHeaderByElements = getSetHeaderByElements;
+
   // reaction
   protected findReactionId = findReactionId;
 
@@ -357,8 +361,8 @@ export class LXCatDatabase {
   public getSetVersionInfo = getSetVersionInfo;
 
   // css/public
+  public listSets = listSets;
   public getSetById = setById;
-  public searchSet = searchSet;
   public getMixtureByIds = byIds;
   public setHistory = setHistory;
   public getItemIdsInSet = getItemIdsInSet;
@@ -366,12 +370,6 @@ export class LXCatDatabase {
 
   // css/get-affiliation
   public getSetAffiliation = getSetAffiliation;
-
-  // css/picker
-  public searchFacets = searchFacets;
-  protected searchContributors = searchContributors;
-  public stateChoices = stateChoices;
-  protected tagChoices = tagChoices;
 
   // cs/picker
   public getSearchOptions = getSearchOptions;
